Small improvements
